## Summary
- [ X] I have read CONTRIBUTING.md to understand how to contribute to this repository :)

String concat for char augmentation in upside_down.py and letter_replacement.py is slowing down the runtime. The current approach starts from an empty string and uses "+=" to add every char. Instead, we could turn each word into a list of chars, and change the chars only at necessary indexes. This new approach can prove to be even more efficient when the aug_p is low.

Runtime analysis:

**Contractions**
OLD: 0.00100s
NEW: 0.00040s
60% faster with p=0.3

**ReplaceSimilarChars / ReplaceSimilarUnicodeChars**
OLD: 0.00020s
NEW: 0.00009s
55% faster with p=0.3

**ReplaceUpsideDown**
OLD: 0.00010s
NEW: 0.00010s
40% faster with p=0.3

### Text
```bash
(base) C:\Users\kire5\Desktop\AugLy>python -m unittest discover -s augly/tests/text_tests/ -p "*"
.......................................
----------------------------------------------------------------------
Ran 39 tests in 0.025s

OK
```