@zpapakipos 2021-11-25
@Cloud9c I just marked this as ready for review so I could import it to test that it works internally. Please add the times for each function you're speeding up in this PR and make any other required changes before I will actually approve it. Thanks!